refactor(bugprone): narrowing espliciti (AppConfigManager/ProcessUtils), widening moltiplicazione (Logger), branch-clone unito (AnimeW); config .clang-tidy: header-filter esclude _deps vendored, catch intenzionali riconosciuti da keyword, IgnoredParameterNames per i due int di httpGetWithRetry

+++ src/config/AppConfigManager.cpp
@@ -141,7 +141,7 @@ namespace Config {
 
         // 2. Budget Thread: Burst (85% CPU) vs Background (50% CPU)
         float usageFactor = burstMode ? 0.85f : 0.50f;
-        int targetThreadsBudget = static_cast<int>(totalThreads * usageFactor);
+        int targetThreadsBudget = static_cast<int>(static_cast<float>(totalThreads) * usageFactor);
         if (targetThreadsBudget < 1)
             targetThreadsBudget = 1;
